objdiff v4.2.4 stopped folding vetted relocation-name disagreements out of
match_percent_normalized, which exposed a class of bugs that had been scoring a
perfect 100.0 while referencing the wrong symbol. These are the first seven
closed here.

Every row was verified equal under functionRelocDiffs=none and mismatched under
name_check before being touched; six of the seven had exactly ONE masked reloc
row with the rest of the function already perfect, so the pairing was not in
question. Whole-binary A/B over two worktrees off the same base, full ninja each
side: 8 functions up, 0 down, matched_code +1,564 B.

Two were real defects rather than naming:
  - soundtouch::TDStretch overloads operator new to _MemAlloc(t, 32) and had no
    matching operator delete, so memory from the game heap went back to the
    global deallocator. Its sibling RateTransposer.h already carries the pair.
  - TrackPanel.h declared TrackerDisplayReset() twice, non-const and const. Only
    the const one is ever defined and only the const one exists in the target,
    so we were emitting a call to a function that does not exist.

The DataArray::Node const class has a single root cause, not two: the sites
hand-expanded arr->Node(i).Int(arr) on a non-const DataArray*, where the
original wrote arr->Int(i) and let the inline const member pick the const
overload.

Two negative results are committed as source comments at their call sites, both
verified codegen-neutral by a 0-up/0-down report A/B: rindex const does fix all
four rows and costs 0.78 canonical points by reshuffling the unmatched rest of
the function, and hex_decode's get_pointer is private on basic_string, reachable
only by removing a self-described fakematch crutch that costs a net matched
function elsewhere. Neither should be reapplied alone.

+++ src/band3/meta_band/SavedSetlist.cpp
@@ -175,7 +175,7 @@ Symbol NetSavedSetlist::GetIdentifyingToken() const {
         MILO_FAIL("Battles should have their own identifying token!\n");
         break;
     default:
-        MILO_FAIL("Bad NetSavedSetlist SetlistType %i!\n", mSetlistType);
+        MILO_FAIL("Bad NetSavedSetlist SetlistType %i!\n", (int)mSetlistType);
         break;
     }
     return gNullStr;
